(PC-34823) refactor(reactions): queries for /features/reactions
