+++ src/features/bookings/pages/Bookings/Bookings.web.test.tsx
@@ -9,7 +9,7 @@ import {
   SubcategoryIdEnum,
 } from 'api/gen'
 import { bookingsSnap } from 'features/bookings/fixtures/bookingsSnap'
-import { useAvailableReaction } from 'features/reactions/api/useAvailableReaction'
+import { useAvailableReactionQuery } from 'features/reactions/queries/useAvailableReactionQuery'
 import { setFeatureFlags } from 'libs/firebase/firestore/featureFlags/__tests__/setFeatureFlags'
 import { useSubcategories } from 'libs/subcategories/useSubcategories'
 import * as bookingsAPI from 'queries/bookings/useBookingsQuery'
@@ -58,8 +58,8 @@ mockUseCategoryIdMapping.mockReturnValue({
   [SubcategoryIdEnum.EVENEMENT_PATRIMOINE]: CategoryIdEnum.MUSEE,
 })
 
-jest.mock('features/reactions/api/useAvailableReaction')
-const mockUseAvailableReaction = useAvailableReaction as jest.Mock
+jest.mock('features/reactions/queries/useAvailableReactionQuery')
+const mockUseAvailableReaction = useAvailableReactionQuery as jest.Mock
 mockUseAvailableReaction.mockReturnValue({
   data: { numberOfReactableBookings: 0, bookings: [] },
 })
