+++ src/features/home/components/IncomingReactionModalContainer/IncomingReactionModalContainer.native.test.tsx
@@ -15,7 +15,7 @@ jest.mock('libs/firebase/remoteConfig/remoteConfig.services')
 jest.mock('libs/firebase/analytics/analytics')
 
 const mockMutate = jest.fn()
-jest.mock('features/reactions/api/useReactionMutation', () => ({
+jest.mock('features/reactions/queries/useReactionMutation', () => ({
   useReactionMutation: () => ({ mutate: mockMutate }),
 }))
