+++ src/features/navigation/helpers/useTabBarItemBadges.ts
@@ -1,10 +1,10 @@
 import { useMemo } from 'react'
 
 import { TabParamList } from 'features/navigation/TabBar/types'
-import { useAvailableReaction } from 'features/reactions/api/useAvailableReaction'
+import { useAvailableReactionQuery } from 'features/reactions/queries/useAvailableReactionQuery'
 
 export const useTabBarItemBadges = (): Partial<Record<keyof TabParamList, number>> => {
-  const { data: availableReactions } = useAvailableReaction()
+  const { data: availableReactions } = useAvailableReactionQuery()
   const numberOfReactableBookings = availableReactions?.numberOfReactableBookings
 
   const routeBadgeMap = useMemo(
